# docs/swagger.yaml
basePath: /api
definitions:
...
  auth.SLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for DAPEN management of employee retirement funds
  title: DAPEN System API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the password of the currently authenticated user
      parameters:
      - description: Old and New Password
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change User Password
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: User Login
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Rotate JWT tokens using a valid refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh JWT Token
      tags:
      - Authentication
...
  /customers:
    get:
      description: Fetch customer data with pagination and search
      parameters:
      - default: 0
        description: Customer type (legacy DBCUSTSUPP.JENIS)
        in: query
        name: customerType
        type: integer
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Customers
      tags:
      - Filter
...
  /me:
    get:
      description: Retrieves the currently authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - User
...
  /perkiraan:
    get:
      description: Fetch perkiraan data with pagination and search
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Comma-separated perkiraan codes to exclude
        in: query
        name: without
        type: string
      - description: Restrict to hutang/piut entries
        in: query
        name: posthutpiut
        type: string
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Perkiraan
      tags:
      - Filter
  /perkiraan/kelompok-kas:
    get:
      description: Fetch kelompok kas data with pagination and search
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Kelompok Kas
      tags:
      - Filter
swagger: "2.0"